Move AccentColorConverter in Avalonia.Controls.Primitives.Converters namespace
This better hides these special-purpose converters
