+++ src/Avalonia.Controls.ColorPicker/ColorPreviewer/ColorPreviewer.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Globalization;
 using Avalonia.Controls.Metadata;
+using Avalonia.Controls.Primitives.Converters;
 using Avalonia.Input;
 using Avalonia.Media;
