+++ src/Avalonia.Controls.ColorPicker/Themes/Fluent/ColorPreviewer.xaml
@@ -1,10 +1,10 @@
 ﻿<Styles xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:primitives="using:Avalonia.Controls.Primitives"
+        xmlns:pc="using:Avalonia.Controls.Primitives.Converters"
         x:CompileBindings="True">
 
   <Styles.Resources>
-    <primitives:AccentColorConverter x:Key="AccentColor" />
+    <pc:AccentColorConverter x:Key="AccentColor" />
   </Styles.Resources>
 
   <Style Selector="ColorPreviewer">
